pyproject.toml
[tool.poetry]
name = "argos-backend"
version = "0.1.0"
description = "Argos Strategic OKR Platform Backend"
authors = ["Antigravity <antigravity@google.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = ">=4.2.0"
djangorestframework = ">=3.14.0"
djangorestframework-simplejwt = ">=5.2.0"
django-cors-headers = ">=4.0.0"
django-filter = ">=23.2"
weasyprint = ">=59.0"
psycopg2-binary = ">=2.9.6"
gunicorn = ">=20.1.0"
python-decouple = ">=3.8"
requests = ">=2.31.0"
whitenoise = ">=6.5.0"
django-encrypted-model-fields = "^0.6.5"
drf-spectacular = "^0.26.0"
django-extensions = "^3.2.3"
python-dotenv = "^1.0.0"
django-storages = {extras = ["google"], version = ">=9.1.0"}
google-cloud-storage = ">=2.10.0"
bleach = ">=6.0.0"
djangorestframework-ratelimit = ">=0.8.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
